requirements.txt
flask
flask-sqlalchemy
flask-cors
flask-jwt-extended
flask-socketio
werkzeug
pyngrok
python-dotenv
pillow
bleach
emoji
profanity-check
textblob
flask-limiter
redis
eventlet